+++ react-app/src/features/common/Header.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import AppBar from '@mui/material/AppBar';
+import Toolbar from '@mui/material/Toolbar';
+import Typography from '@mui/material/Typography'
+
+const Header = () => {
+    return (
+        <div>
+            <AppBar position="static">
+                <Toolbar>
+                    <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
+                        Blog
+                    </Typography>
+                </Toolbar>
+            </AppBar>
+        </div>
+    );
+};
+
+export default Header;
